cleaning up unneccessary prop drilling
